[bug/V5.3.0/Android/825_delete-alert] Redmine #825: [IT] Unable to delete from Print Job History after rapid taps
Cause:
PrintJobsFragment.showDeleteDialog() checks if the delete dialog is currently showing. The condition used is `if mConfirmDialog != null`.

When the dialog is dismissed via OK or Cancel button press, Android sends a onCancel() or onConfirm() message to PrintJobsFragment, where mConfirmDialog will be set to null.

However, if the dialog is dismissed due to multiple rapid taps, the onCancel() or onConfirm() methods are not called, and mConfirmDialog is still not null. This means that PrintJobsFragment continues to assume that the dialog is still showing, and will not show the dialog again.

It seems that in the latest update of FireOS, Android calls the onCancel() method when the dialog is dismissed from rapid taps, so the issue no longer occurred there.

Solution:
Add a isShowing() function to the confirmation dialog.
When PrintJobsFragment.showDeleteDialog() checks if the delete dialog is currently showing, check if the alert dialog is actually showing.